chore: 폴더 위치 변경(추후 재 변경 가능성 매우 큽니다)
